requirements.txt
torch>=2.0
torchaudio>=2.0
torchvision>=0.15
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
scipy>=1.11
faiss-cpu>=1.7
